refactor: reorganize tests project
