+++ apps/api/Tests/Receipt/Receipt.Tests/Receipt.Tests.csproj
@@ -22,4 +22,8 @@
     <Using Include="Xunit" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Application\" />
+  </ItemGroup>
+
 </Project>
