+++ apps/api/Tests/Receipt/Receipt.Tests/TestHelpers/ReceiptTestFactory.cs
@@ -1,6 +1,6 @@
 ﻿using Receipt.Domain.Models;
 
-namespace Receipt.Tests;
+namespace Receipt.Tests.TestHelpers;
 
 public static class ReceiptTestFactory
 {
